+++ app/src/main/java/com/codesoom/assignment/application/product/ProductService.java
@@ -1,9 +1,9 @@
-package com.codesoom.assignment.application;
+package com.codesoom.assignment.application.product;
 
 import com.codesoom.assignment.ProductNotFoundException;
-import com.codesoom.assignment.domain.Product;
-import com.codesoom.assignment.domain.ProductRepository;
-import com.codesoom.assignment.dto.ProductData;
+import com.codesoom.assignment.domain.product.Product;
+import com.codesoom.assignment.domain.product.ProductRepository;
+import com.codesoom.assignment.dto.product.ProductData;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
