+++ app/src/main/java/com/codesoom/assignment/controllers/product/ProductController.java
@@ -1,8 +1,8 @@
-package com.codesoom.assignment.controllers;
+package com.codesoom.assignment.controllers.product;
 
-import com.codesoom.assignment.application.ProductService;
-import com.codesoom.assignment.domain.Product;
-import com.codesoom.assignment.dto.ProductData;
+import com.codesoom.assignment.application.product.ProductService;
+import com.codesoom.assignment.domain.product.Product;
+import com.codesoom.assignment.dto.product.ProductData;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
