+++ app/src/main/java/com/codesoom/assignment/domain/product/Product.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.domain;
+package com.codesoom.assignment.domain.product;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
